Rename file to match compnent
